Add files via upload
This password generator creates a random password including numbers, upper/lowercase letters, and symbols. The password length is also semi-random depending on desired strength. 

It works, however after looking at others' solutions, I realized the code could be a lot more efficient. Much of the code could be optimized to one or just a few lines and still have the same capability; namely, randLetter + randNum + randChara could all be combined into the same string. That would eliminate the 3 "if-else's" and save on code.